fix: added link to logo and moved the component
