+++ src/features/app/organisms/TheHeader.js
@@ -1,16 +1,20 @@
 import React from "react";
+import { Link } from "react-router-dom";
 
-import Logo from "../atoms/Logo";
-import TheHeaderNav from "../molecules/TheHeaderNav";
+import Divider from "components/Divider";
+import Logo from "components/Logo";
 import { ReactComponent as TmdbLogo } from "images/tmdb-logo.svg";
+import TheHeaderNav from "../molecules/TheHeaderNav";
 
 export default function TheHeader() {
   return (
-    <header className="h-16 bg-gray-800 flex items-center justify-between px-10 shadow-md rounded">
+    <header className="h-16 bg-gray-800 flex items-center justify-between px-10 shadow-md rounded z-10 mt-5 mx-5">
       <div className="flex items-center h-full py-3">
-        <Logo />
+        <Link to="/">
+          <Logo />
+        </Link>
 
-        <hr className="border-l h-full border-gray-700 mx-10" />
+        <Divider vertical className="mx-10" />
 
         <TmdbLogo className="w-32" />
       </div>
